pyproject.toml
[tool.poetry]
name = "auto-suggestion-programm"
version = "0.1.0"
description = "Suggest few plate numbers in case if we got wrong one."
authors = ["NecronWasTaken <joleyjoker14@gmail.com>"]
readme = "README.md"
packages = [{include = "auto_suggestion_programm"}]

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
thefuzz = "^0.20.0"
pillow = "^10.0.0"
numpy = "^1.25.2"
opencv-python = "^4.8.0.76"
scipy = "^1.11.2"
scikit-image = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
